requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
azure-storage-blob==12.8.1
